Migrate StateMachine clock from Duration to LocalTimestamp.

+++ crates/node/src/io_loop/actions.rs
@@ -599,7 +599,7 @@ where
         // by `source` so dashboards can separate the three commit paths
         // (aggregator/header/sync), which have materially different latencies
         // under the 2-chain rule.
-        let now_ms = self.state.now().as_millis() as u64;
+        let now_ms = self.state.now().as_millis();
         let commit_latency_secs =
             (now_ms.saturating_sub(commit.block.header().timestamp.as_millis())) as f64 / 1000.0;
         metrics::record_block_committed(height.0, commit_latency_secs, commit.source.as_str());
